+++ src/main/java/io/discloader/discloader/core/entity/user/UserAvatar.java
@@ -4,15 +4,15 @@ import java.net.MalformedURLException;
 import java.net.URL;
 
 import io.discloader.discloader.entity.IIcon;
-import io.discloader.discloader.util.DLUtil.Endpoints;
+import io.discloader.discloader.network.util.Endpoints;
 
 public class UserAvatar implements IIcon {
 
 	private String hash;
 	private long userID;
-	private int discrim;
+	private short discrim;
 
-	public UserAvatar(String hash, long userID, int discrim) {
+	public UserAvatar(String hash, long userID, short discrim) {
 		this.hash = hash;
 		this.userID = userID;
 	}
