+++ src/main/java/io/discloader/discloader/entity/user/IUser.java
@@ -32,7 +32,7 @@ public interface IUser extends ISnowflake, ICreationTime {
 	/**
 	 * @return
 	 */
-	int getDiscriminator();
+	short getDiscriminator();
 
 	/**
 	 * @return
